Update Driver.java constants and indentation
I fixed the implementation of the driver reading in user declared constants from files. In java, args[0] is not the program name but instead the first argument so that had to be adjusted. Also many constants require spaces so I had to change line.split()'s delimiter to ',' instead. Also I changed the indentation.